Refactor folders and add support for mingw and cygwin
