+++ CMakeLists.txt
@@ -1,5 +1,7 @@
 cmake_minimum_required(VERSION 2.6)
 project (nesmu)
 
-add_subdirectory(libs)
+add_subdirectory(core)
+add_subdirectory(reader)
+add_subdirectory(donut)
 add_subdirectory(emulator)
