+++ core/include/core/macro.hpp
