+++ donut/include/donut/flags.hpp
@@ -1,0 +1,18 @@
+//
+// Created by jalal on 16/01/16.
+//
+
+#ifndef DONUT_FLAGS_HPP
+#define DONUT_FLAGS_HPP
+
+# include "donut/exports.hpp"
+
+namespace donut
+{
+    struct EXPORT_DLL flags
+    {
+        static bool DEBUG; // need to be activated by the project other flags will come later
+    };
+}
+
+#endif //DONUT_FLAGS_HPP
