+++ emulator/include/emulator/constants.hpp
