+++ emulator/include/emulator/memory.hpp
@@ -1,0 +1,16 @@
+//
+// Created by jalal on 16/01/16.
+//
+
+#ifndef NESMU_MEMORY_HPP
+#define NESMU_MEMORY_HPP
+
+# include <array>
+# include "emulator/constants.hpp"
+
+namespace nesmu
+{
+    using Memory = std::array<uint8_t, MEM_SIZE>;
+}
+
+#endif //NESMU_MEMORY_HPP
