--- libs/CMakeLists.txt
@@ -1,6 +1,0 @@
-cmake_minimum_required(VERSION 2.6)
-project(nesmu_libs)
-
-add_subdirectory(core)
-add_subdirectory(reader)
-add_subdirectory(donut)
