--- libs/core/include/types.hpp
@@ -1,18 +1,0 @@
-//
-// Created by Chaabane on 1/13/16.
-//
-
-#ifndef NESMU_TYPES_HPP
-#define NESMU_TYPES_HPP
-
-# include <vector>
-
-namespace nesmu
-{
-    namespace core
-    {
-        using ByteVector = std::vector<char>;
-    }
-}
-
-#endif //NESMU_TYPES_HPP
