--- libs/donut/include/donut/constants.hpp
@@ -1,14 +1,0 @@
-//
-// Created by jalal on 16/01/16.
-//
-
-#ifndef DONUT_CONSTANTS_H
-#define DONUT_CONSTANTS_H
-
-# ifdef WIN32
-#  define IS_WINDOWS 1
-# else
-#  define IS_WINDOWS 0
-# endif
-
-#endif // DONUT_CONSTANTS_H
