--- libs/donut/include/donut/exports.hpp
@@ -1,15 +1,0 @@
-//
-// Created by jalal on 16/01/16.
-//
-
-#ifndef DONUT_EXPORTS_HPP
-#define DONUT_EXPORTS_HPP
-
-
-# ifdef WIN32
-#  define EXPORT_DLL __declspec(dllexport)
-# else
-#  define EXPORT_DLL
-# endif
-
-#endif //DONUT_EXPORTS_HPP
