--- libs/donut/include/donut/flags.hpp
@@ -1,17 +1,0 @@
-//
-// Created by jalal on 16/01/16.
-//
-
-#ifndef DONUT_FLAGS_HPP
-#define DONUT_FLAGS_HPP
-
-namespace donut
-{
-    struct flags
-    {
-        static bool DEBUG; // need to be activated by the project other flags will come later
-
-    };
-}
-
-#endif //DONUT_FLAGS_HPP
